requirements.txt
gitpython
pytest
dogpile.cache==0.6.4
docker~=3.0
pytest-flake8
pytest-cov
pytest-mock
requests
python-vagrant
fabric3
entrypoints>=0.2.3

